git-gui: schedule rescan after actions triggered by mouse clicks
A mouse click on a UI element of Git GUI should trigger that
action. This does not always happen, in particluar, when the same
mouse click also activiates the Git GUI main window. In this case,
the auto-rescan is triggered first, which as a side-effect
inhibits many operations that should have happened otherwise.

Do not let the auto-rescan have that much priority. For example,
when an entry in the file list is clicked, that entry's diff should
be computed. But it isn't, because the rescan has taken over and
clogs the command execution.

Signed-off-by: Johannes Sixt <j6t@kdbg.org>