git-gui: use git-clone
git-gui clones a repository by invoking git-plumbing commands, in proc
do_clone, rather than using git-clone.  The justification was that the
low-level commands are guaranteed to provide a stable interface, while
the higher level commands such as git-clone may not be stable. This
approach requires git-gui to continually evolve by mirroring new
features in git itself, which has not happened, while the user interface
in git-clone has proven very stable. Also, git-gui does directly call
many other non-plumbing commands in git's repertoire.

do_clone's last significant functionality change was in 2015, and
updates are required for shallow clones, the reftable backend, and
perhaps other features and bugs. For instance, I had reports of git-gui
failing to correctly clone repositories prior to 2015, resulting in
essentially the patch given here. There were very minor changes needed
to bring this up to date, all regarding internal git-gui interfaces and
not git-clone itself. So, the interface to git-clone is clearly stable
enough to not be a problem.

Supporting new use-cases with this requires exposing new options in the
clone dialog, then passing flags to git-clone. This avoids updating
do_clone to understand those options, reducing the maintenance burdens.

So, teach git-gui to use git-clone.  This change is in one patch as
there is no obvious incremental path to migration. The existing dialog /
options / status screen are unchanged, so no new use-cases are exposed.
The status messages shown are now what git-clone provides, which is the
known user-visible change with this patch.

Signed-off-by: Mark Levedahl <mlevedahl@gmail.com>