git-gui: use create-shortcut for Windows desktop shortcuts
git-gui since 51a41ac4ef uses a combination of Tcl, javascript, and the
Windows provided wscript.exe to create shortcuts on the Windows desktop.
This execution chain requires encoding all arguments on a command line,
using quotation marks and escape characters. The relatively simplistic
encoding used (add double quotes around any argument with a space in it)
does not handle all legal path-names: as a result, the created shortcut
is not always usable.

Git for Windows, based upon MSYS, includes 'create-shortcut.exe' to
create desktop shortcuts.  This utility accepts all arguments in an argv
array, and directly invokes the relevant win32 api functions to create a
shortcut in similar manner, so has no need to re-encode a command line
with escape characters and quotations. This works in many cases where
the current approach fails.  So, let's switch to using create-shortcut
on Windows.

Signed-off-by: Mark Levedahl <mlevedahl@gmail.com>