git-gui: use create-shortcut for Windows desktop shortcuts
git-gui since 51a41ac4ef uses a combination of Tcl, javascript, and the
Windows provided wscript.exe to create shortcuts on the Windows desktop.
That chain of language and script host is unique to this particular use,
while git for windows' underlying distribution provides and maintains
'create-shortcut.exe' for this same purpose.  Let's switch to that,
allowing us to simplify git-gui itself.

Signed-off-by: Mark Levedahl <mlevedahl@gmail.com>