git-gui: use /cmd/git-gui.exe for shortcut
git-gui on Windows offers to create a desktop shortcut file that starts
git-gui on a given repository. As this shortcut file is executed from a
pure Windows environment, it is important that all environment variables
are correctly set as they may not set in the global or user Windows
environment.  Git for Windows (g4w) provides wrapper functions to launch
git-gui, bash, and others for this exact reason [1]. However, g4w does
not install the git-gui.exe wrapper in a directory on PATH, so git-gui
does not find this wrapper when creating shortcuts. git-gui created
shortcuts therefore may not work correctly.

Fix this by first looking for g4w's git-gui wrapper, which is
'/cmd/git-gui.exe' in bash's POSIX path-name space. If not found, fall
back to the current approach. Note: this fall back is questionable as
the only environment known to work is g4w's.  Change the variable name
for this from 'cmdLine' to 'link_target', as the latter is much more
descriptive of the use.

[1] https://gitforwindows.org/git-wrapper.html

Signed-off-by: Mark Levedahl <mlevedahl@gmail.com>