git-gui: Revert "git-gui - use git-hook, honor core.hooksPath"
The recent changes that protect calls of Tcl's `open` and `exec` from
undesirable redirection operations have been backported to Git releases
before 2.36.0. If a patch with these changes is applied to such an old
version, it generates a conflict in 'githook_read', because the newer
versions use `git hook` to invoke a hook rather than to invoke it
directly. In this case, the resolution is very non-trivial.

Undo the change that introduced the use of `git hook`, 0730a5a3a5e6,
and substitute the correct resolution of the ensuing conflict. The
intent of this commit is to help maintainers of distribution to
backport the security fixes correctly.

The original code had a special path for Windows, but we do not
resurrect it here, because the prevailing distribution is
Git-for-Windows, which is based on a modern version of Git and need
not be helped.

Signed-off-by: Johannes Sixt <j6t@kdbg.org>